Fold metadatafetcher.py and extentreport.py into create_reports.py
