--- app/extentreport.py
@@ -1,386 +1,0 @@
-import os
-import boto3
-import json
-import xlsxwriter
-import humanize
-import requests
-from urllib.parse import urlparse
-import shutil
-
-import settings
-
-MD5S = []
-
-def report(campus, version):
-    '''
-    for a given campus:
-        - get metadata files for campus from S3
-        - parse out file stats metadata
-        - create spreadsheet of stats
-    '''
-
-    # create the excel workbook
-    tmp_dir = settings.TEMP
-    if not os.path.exists(tmp_dir):
-        os.makedirs(tmp_dir)
-    outfile = f"{campus}-extent-stats-{version}.xlsx"
-    outpath = os.path.join(tmp_dir, outfile)
-    workbook = xlsxwriter.Workbook(outpath)
-    bold_format = workbook.add_format({'bold': True})
-    summary_worksheet = workbook.add_worksheet('Summary')
-
-    # write the headings
-    headings = [
-        "Project Folder",
-        "Doc Count",
-        "Unique Main File Count",
-        "Main File Size",
-        "Unique Files Tab Count",
-        "Files Tab Count",
-        "Unique Aux File Count",
-        "Aux File Size",
-        "Unique Derivative File Count",
-        "Derivative File Size",
-        "Total Unique File Count",
-        "Total File Size"
-    ]
-    row = 0
-    col = 0
-    for h in (headings):
-        summary_worksheet.write_string(0, col, h, bold_format)
-        summary_worksheet.set_column(col, col, len(h))
-        col = col + 1
-
-    # increment
-    row += 1
-
-    # create a file to contain a list of all docs for QA purposes
-    doclist_file = f"{campus}-doclist-{version}.txt"
-    doclist_path = os.path.join(tmp_dir, doclist_file)
-    if os.path.exists(doclist_path):
-        os.remove(doclist_path)
-
-    summary_doc_count = 0
-    summary_stats = {
-        "main_count": 0,
-        "main_size": 0,
-        "filetab_count": 0,
-        "filetab_size": 0,
-        "aux_count": 0,
-        "aux_size": 0,
-        "deriv_count": 0,
-        "deriv_size": 0,
-        "total_count": 0,
-        "total_size": 0
-    }
-
-    prefixes = get_child_prefixes(campus, version)
-
-    for prefix in prefixes:
-        print(f"Aggregating stats for {prefix}")
-        stats = get_stats(campus, version, prefix)
-
-        rowname = prefix.split('/')[-1]
-        write_stats(stats, summary_worksheet, row, rowname)
-        row += 1
-
-        # write doc info to a file
-        if not os.path.exists(doclist_path):
-            with open(doclist_path, "w") as f:
-                for doc in stats['docs']:
-                    f.write(f"{doc}")
-        else:
-            with open(doclist_path, "a") as f:
-                for doc in stats['docs']:
-                    f.write(f"{doc}")
-
-        summary_doc_count += stats['doc_count']
-        summary_stats['main_count'] += stats['main_count']
-        summary_stats['main_size'] += stats['main_size']
-        summary_stats['filetab_count'] += stats['filetab_count']
-        summary_stats['filetab_size'] += stats['filetab_size']
-        summary_stats['aux_count'] += stats['aux_count']
-        summary_stats['aux_size'] += stats['aux_size']
-        summary_stats['deriv_count'] += stats['deriv_count']
-        summary_stats['deriv_size'] += stats['deriv_size']
-        summary_stats['total_count'] += stats['total_count']
-        summary_stats['total_size'] += stats['total_size']
-
-    summary_stats['doc_count'] = summary_doc_count
-
-    rowname = 'TOTALS'
-    write_stats(summary_stats, summary_worksheet, row, rowname)
-
-    workbook.close()
-
-    # load files to S3
-    data_loc = urlparse(settings.REPORTS)
-    if data_loc.scheme == 's3':
-        bucket = data_loc.netloc
-        key = f"{data_loc.path}/{campus}/{version}/{outfile}".lstrip('/')
-        load_to_s3(bucket, key, outpath)
-        key = f"{data_loc.path}/{campus}/{version}/{doclist_file}".lstrip('/')
-        load_to_s3(bucket, key, doclist_path)
-    elif data_loc.scheme == 'file':
-        dest_dir = os.path.join(data_loc.path, campus, version)
-        if not os.path.exists(dest_dir):
-            os.makedirs(dest_dir)
-        destination = os.path.join(dest_dir, outfile)
-        print(f"Writing file://{destination}")
-        shutil.copyfile(outpath, destination)
-
-        destination = os.path.join(dest_dir, doclist_file)
-        print(f"Writing file://{destination}")
-        shutil.copyfile(doclist_path, destination)
-
-    # delete tmp files
-    os.remove(outpath)
-    os.remove(doclist_path)
-
-def get_child_prefixes(campus, version):
-    data_loc = urlparse(settings.METADATA)
-    if data_loc.scheme == 'file':
-        child_prefixes = []
-        path = os.path.join(data_loc.path, campus, version)
-        dir_list = os.listdir(path)
-        return dir_list
-    else:
-        s3_client = boto3.client('s3')
-        paginator = s3_client.get_paginator('list_objects_v2')
-        bucket = data_loc.netloc
-        prefix = data_loc.path
-        prefix = prefix.lstrip('/')
-        prefix = f"{prefix}/{campus}/{version}"
-        pages = paginator.paginate(
-            Bucket=bucket,
-            Prefix=prefix
-        )
-
-        child_prefixes = []
-        folder_prefix_parts_count = len(prefix.split('/'))
-        for page in pages:
-            for item in page['Contents']:
-                parts = item['Key'].split('/')
-                child_prefix = '/'.join(parts[0:folder_prefix_parts_count + 1])
-                if not child_prefix in child_prefixes:
-                    child_prefixes.append(child_prefix)
-
-    return child_prefixes
-
-def get_stats(campus, version, prefix):
-
-    doc_count = 0
-
-    stats = {
-        "main_count": 0,
-        "main_size": 0,
-        "filetab_count": 0,
-        "filetab_size": 0,
-        "aux_count": 0,
-        "aux_size": 0,
-        "deriv_count": 0,
-        "deriv_size": 0,
-        "total_count": 0,
-        "total_size": 0,
-        "docs": []
-    }
-
-    data_loc = urlparse(settings.METADATA)
-    if data_loc.scheme == 'file':
-        metadata_dir = os.path.join(data_loc.path, campus, version, prefix)
-        for file in os.listdir(metadata_dir):
-            with open(os.path.join(metadata_dir, file), "r") as f:
-                for line in f.readlines():
-                    stats = add_doc_to_stats(stats, line)
-                    doc_count += 1
-    elif data_loc.scheme == 's3':
-        s3_client = boto3.client('s3')
-        paginator = s3_client.get_paginator('list_objects_v2')
-        pages = paginator.paginate(
-            Bucket=data_loc.netloc,
-            Prefix=prefix
-        )
-        for page in pages:
-            for item in page['Contents']:
-                #print(f"getting s3 object: {item['Key']}")
-                response = s3_client.get_object(
-                    Bucket=data_loc.netloc,
-                    Key=item['Key']
-                )
-                for line in response['Body'].iter_lines():
-                    stats = add_doc_to_stats(stats, line)
-                    doc_count += 1
-    else:
-        raise Exception(f"Unknown data scheme: {data_loc.scheme}")
-
-    # Total Items (including components of complex objects; some may not have associated files)
-    stats['doc_count'] = doc_count
-
-    return stats
-
-def add_doc_to_stats(stats, doc):
-
-    doc = json.loads(doc)
-    doc_extent = get_extent(doc)
-
-    stats['docs'].append(f"{doc['uid']}, {doc['path']}\n")
-    stats['main_count'] += doc_extent['main_count']
-    stats['main_size'] += doc_extent['main_size']
-    stats['filetab_count'] += doc_extent['filetab_count']
-    stats['filetab_size'] += doc_extent['filetab_size']
-    stats['aux_count'] += doc_extent['aux_count']
-    stats['aux_size'] += doc_extent['aux_size']
-    stats['deriv_count'] += doc_extent['deriv_count']
-    stats['deriv_size'] += doc_extent['deriv_size']
-    stats['total_count'] += doc_extent['total_count']
-    stats['total_size'] += doc_extent['total_size']
-
-    return stats
-
-def get_extent(doc):
-    # query db for each record as a workaround while ES API endpoint is broken
-    if settings.NUXEO_API_ES_ENDPOINT_BROKEN:
-        uid = doc['uid']
-        doc = get_metadata_from_db(uid)
-
-    extent = {
-        "main_count": 0,
-        "main_size": 0,
-        "filetab_count": 0,
-        "filetab_size": 0,
-        "aux_count": 0,
-        "aux_size": 0,
-        "deriv_count": 0,
-        "deriv_size": 0,
-        "total_count": 0,
-        "total_size": 0
-    }
-
-    properties = doc['properties']
-
-    if properties.get('file:content'):
-        content = properties.get('file:content')
-        if not content['digest'] in MD5S:
-            MD5S.append(content['digest'])
-            extent['main_count'] += 1
-            extent['main_size'] += int(content['length'])
-            #print(f"main {extent['main_count']} file:content {content['name']} {int(content['length'])}")
-
-    # Original files vs file:content?
-    if properties.get('picture:views'):
-        for view in properties.get('picture:views'):
-            content = view['content']
-            if not content['digest'] in MD5S:
-                MD5S.append(content['digest'])
-                extent['deriv_count'] += 1
-                extent['deriv_size'] += int(content['length'])
-                #print(f"deriv {extent['deriv_count']} picture:views {content['name']} {view['description']} {int(content['length'])}")
-
-    # extra_files:file
-    if properties.get('extra_files:file'):
-        file = properties.get('extra_files:file')
-        for f in file:
-            if f.get('blob') and not f['blob']['digest'] in MD5S:
-                blob = f.get('blob')
-                MD5S.append(blob['digest'])
-                extent['aux_count'] += 1
-                extent['aux_size'] += int(blob['length'])
-                #print(f"aux {extent['aux_count']} extra_files {blob['name']} {int(blob['length'])}")
-
-    # files:files
-    if properties.get('files:files'):
-        files = properties.get('files:files')
-        for file in files:
-            if file.get('file') and not file['file']['digest'] in MD5S:
-                file = file.get('file')
-                extent['filetab_count'] += 1
-                extent['filetab_size'] += int(file['length'])
-                #print(f"filetab {extent['filetab_count']} files:files {file['name']} {int(file['length'])}")
-
-    # vid:storyboard
-    if properties.get('vid:storyboard'):
-        storyboard = properties.get('vid:storyboard')
-        for board in storyboard:
-            if board.get('content') and not board['content']['digest'] in MD5S:
-                content = board.get('content')
-                extent['deriv_count'] += 1
-                extent['deriv_size'] += int(content['length'])
-                #print(f"deriv {extent['deriv_count']} storyboard {content['name']} {int(content['length'])}")
-
-    # vid:transcodedVideos
-    if properties.get('vid:transcodedVideos'):
-        videos = properties.get('vid:transcodedVideos')
-        for vid in videos:
-            if vid.get('content') and not vid['content']['digest'] in MD5S:
-                content = vid.get('content')
-                extent['deriv_count'] += 1
-                extent['deriv_size'] += int(content['length'])
-                #print(f"deriv {extent['deriv_count']} vid:transcodedVideos {content['name']} {int(content['length'])}")
-
-    # auxiliary_files:file
-    if properties.get('auxiliary_files:file'):
-        auxfiles = properties.get('auxiliary_files:file')
-        for af in auxfiles:
-            if af.get('content') and not af['content']['digest'] in MD5S:
-                content = af.get('content')
-                extent['deriv_count'] += 1
-                extent['deriv_size'] += int(content['length'])
-
-    # 3D
-    if properties.get('threed:transmissionFormats'):
-        formats = properties.get('threed:transmissionFormats')
-        for format in formats:
-            if format.get('content') and not format['content']['digest'] in MD5S:
-                content = format.get('content')
-                extent['deriv_count'] += 1
-                extent['deriv_size'] += int(content['length'])
-
-    extent['total_count'] = extent['main_count'] + extent['filetab_count'] + extent['deriv_count'] + extent['aux_count']
-    extent['total_size'] = extent['main_size'] + extent['filetab_size'] + extent['deriv_size'] + extent['aux_size']
-
-    return extent
-
-def write_stats(stats, worksheet, rownum, rowname):
-
-    formatted_data = [
-        rowname,
-        stats['doc_count'],
-        stats['main_count'],
-        humanize.naturalsize(stats['main_size'], binary=True),
-        stats['filetab_count'],
-        humanize.naturalsize(stats['filetab_size'], binary=True),
-        stats['aux_count'],
-        humanize.naturalsize(stats['aux_size'], binary=True),
-        stats['deriv_count'],
-        humanize.naturalsize(stats['deriv_size'], binary=True),
-        stats['total_count'],
-        humanize.naturalsize(stats['total_size'], binary=True)
-    ]
-
-    col = 0
-    for d in (formatted_data):
-        worksheet.write(rownum, col, d)
-        col = col + 1
-
-def get_metadata_from_db(uid):
-    # GET 'https://nuxeo.cdlib.org/nuxeo/site/api/v1/id/32f09ee0-dcc0-4746-90c4-ba4c710447cd' -H 'X-NXproperties: *' -H 'X-NXRepository: default' -H 'content-type: application/json' -u Administrator -p
-    url = u'/'.join([settings.NUXEO_API, "id", uid])
-    request = {'url': url, 'headers': settings.NUXEO_REQUEST_HEADERS}
-    response = requests.get(**request)
-    response.raise_for_status()
-    json_resp = response.json()
-    return json_resp
-
-def load_to_s3(bucket, key, filepath):
-    s3_client = boto3.client('s3')
-
-    print(f"Writing s3://{bucket}/{key}")
-    try:
-        response = s3_client.upload_file(
-            Filename=filepath,
-            Bucket=bucket,
-            Key=key
-        )
-    except Exception as e:
-            print(f"ERROR loading to S3: {e}")
-
